test/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-RESTful==0.3.9
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
pyasn1==0.4.8
pycryptodome==3.14.1
pymongo==4.0.1
pytz==2021.3
six==1.16.0
urllib3==1.26.8
Werkzeug==2.0.3
